+++ R/with_progress.R
@@ -23,6 +23,8 @@
 #' report on the current amount progress when the evaluation was terminated
 #' by the interrupt, e.g. when a user pressed Ctrl-C in an interactive session,
 #' or a batch process was interrupted because it ran out of time.
+#' Note that it's optional for a progress handler to support this and only
+#' some do.
 #'
 #' @param interval (numeric) The minimum time (in seconds) between
 #' successive progression updates from handlers.